+++ tests/testthat/test-addPareArticle.R
@@ -1,0 +1,7 @@
+test_that("minimal", {
+  suppressWarnings(addPareArticle(repo))
+
+  path <- file.path(repo$getPath(), "vignettes", "articles", "PareReport.Rmd")
+
+  expect_true(file.exists(path))
+})