+++ tests/testthat/test-getDoCallFromLines.R
@@ -1,0 +1,9 @@
+lines <- c(
+  "do.call(mean, list(c(1, 2, 3)))",
+  "do.call('mean', list(c(1, 2, 3)))",
+  "do.call(\"mean\", list(c(1, 2, 3)))"
+)
+
+test_that("minimal", {
+  expect_true(all(PaRe:::getDoCallFromLines(lines) == "mean"))
+})