+++ NEWS.md
@@ -3,6 +3,21 @@ PaRe 0.1.10
 1. Fixed issue with pattern matching of files in .gitignore.
 2. Fixed issue with retrieving the correct amount of arguments from a function.
 3. Changed colour highlights for report.
+4. Added addPareArticle function to add a PaRe report as an article.
+5. Added setup for unittesting.
+6. Added unittest for:
+    - getDefaultPermittedPackages
+    - addPareArticle
+    - getApplyCall
+    - getApplyFromLines
+    - getDlplyCall
+    - getDlplyCallFromLines
+    - getDoCall
+    - getDoCallFromLines
+    - getFunCall
+    - getMultiLineFun
+    - checkDependencies
+7. Fixed issue detecting functions called by plyr::dlply
 
 PaRe 0.1.9
 ==========